Make address comparison use size_t not int.
svn path=/trunk/; revision=15390

+++ applications/gaia/tcl/tcl8.0.3/generic/tclObj.c
@@ -557,7 +557,7 @@ TclAllocateFreeObjects()
 {
     Tcl_Obj tmp[2];
     size_t objSizePlusPadding =	/* NB: this assumes byte addressing. */
-	((int)(&(tmp[1])) - (int)(&(tmp[0])));
+	((size_t)(&(tmp[1])) - (size_t)(&(tmp[0])));
     size_t bytesToAlloc = (OBJS_TO_ALLOC_EACH_TIME * objSizePlusPadding);
     char *basePtr;
     register Tcl_Obj *prevPtr, *objPtr;
